+++ server/@types/ui/index.d.ts
@@ -28,7 +28,6 @@ import {
   UserQualification,
   ApprovedPremisesUserRole as UserRole,
 } from '@approved-premises/api'
-import { ApTypeCriteria } from '../../utils/placementCriteriaUtils'
 
 interface TasklistPage {
   body: Record<string, unknown>
@@ -471,15 +470,3 @@ export type DepartureFormSessionData = Partial<
     notes: string
   }
 >
-
-export type SpaceSearchState = {
-  applicationId: string
-  postcode: string
-  apType: ApTypeCriteria
-  apCriteria: Array<SpaceSearchApCriteria>
-  roomCriteria: Array<SpaceSearchRoomCriteria>
-  startDate: string
-  durationDays: number
-  arrivalDate?: string
-  departureDate?: string
-}
